pyproject.toml
[project]
name = "slaude"
version = "0.1.0"
description = "Slack + Claude Code bridge — send coding tasks from Slack, get results back"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt[async]>=1.18.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
slaude = "slaude.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
